requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.14.0
pillow>=10.1.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
pydantic[email]>=2.0.0
